Imported system variables are tracked by FBuild
When "#import VAR" is parsed the system variable VAR is tracked by FBuild.
Tracked variable names and their hash values are written in the FDB files.
At loading the variables will be imported and the new hash values will be compared to previous ones.
Reparsing will occur if these values differ or if the variable was not found in system environment.

This patch increments NODE_GRAPH_CURRENT_VERSION, since it changes the format of FDB files.
A clean build will occur.