New syntax for variables `^SomeVar`
`^SomeVar` will point to the first previous declaration of '.SomeVar' **in a parent scope**.
````
.Array = { 'a', 'b', 'c' }
.All = ''
ForEach( .Item in .Array )
{
    .All = .Item
    ^All + .Item
    Print( .All ) // "a", "b", "c"
    Print( ^All ) // "a", "ab", "abc"
    {
        ^All + .Item
        Print( ^All ) // "aa", "bb", "cc"
    }
}
Print( .All ) // "abc"
````

`^SomeVar` will be unknown if '.SomeVar' was not declared in a parent scope.
````
.SomeVar = 'Foo'
^SomeVar = 'Bar' // FASTBuild Error #1009 - Unknown variable.
````

The parameter 'frame' in methods BFFStackFrame::SetVar*() is now mandatory.
It was safer, we don't want other code to be silently ignoring this change without a review.