Allow concatenation/assigning of empty arrays to other arrays
Following statements are valid when .X is a string but not when it's a
struct:
    .E = {}
    .A = { .X }
    .T1 = {} + .X
    .T2 = {} + { .X }
    .T3 = {} + .A
    .T4 = {} + { .A }
    .T5 = { .X } + {}
    .T6 = { .X } + .E
    .T7 = { .X } + { .E }
This is a side effect of empty arrays being stored as ArrayOfStrings.

This change allows treatment of empty ArrayOfStrings as empty
ArrayOfStructs and vice versa, making those statements valid when .X is
a struct as well.