PreprocessorGenericDependencies
Added support for PreprocessorGenericDependencies option to the
ObjectList node. When enabled the custom preprocessor will be used one
to generate dependency information. The preprocessor must write out the
list of dependencies to stdout.

This flag makes it possible to add support for additional compilers and
tools that introduce dynamic dependencies. For instance, the QT resource
compiler could use this method rather than needing to be handled as a
special case in the code.