UnityNode has an optional "UnityInputIsolateListFile" file parameter
The goal is to generalize "UnityInputIsolateWritableFiles".

The user can provide a list of files filled by a third party tool (git!).

"UnityInputIsolateListFile" must point to that file.
At build time it will be loaded and each line will be considered as a valid filename.
Any file matching this list of filenames it will be isolated from the unity (same behavior than read only).
Changing the list won't necessarly invalidate the fdb since the file beeing parsed at runtime.
The file can be using both LF & CRLF.
Won't fail the build if the file did not exist.