Extend sleep workaround in BFFDrifted test to Linux
Some file systems used on Linux (ext2, ext3, reiserfs) have only a
second resolution for timestamps. And even file systems with a higher
timestamp resolution can in practice be limited to a millisecond
resolution because of time caching in the kernel (see current_fs_time()
in the kernel).
Thus to avoid spurious failures we should sleep for 1s before clearing
fbuild.bff file in the BFFDrifted test.