Fix one more case of partially copied AString
See detailed reasoning in e88b121fb5ccaeba0728417a27695b8ac18e2d40.

This case was found with libFuzzer and MemorySanitizer on an input with
a null character embedded in BFF string. Minimal reproducer is (5th
character should be '\0' instead of '0'):
   .S='01234567'+.S+.S+.S+.S+.S+.S+.S+.S
   +'X'
   -' '
In this case the first line creates a string of length 2048 which is the
maximum that can be stored inside variable 'value' in
BFFParser::StoreVariableString. Also this string has '\0' as the first
character.
Second line triggers allocation of a new buffer on the heap, and
AString::Grow copies data only up to the first null.
Third line scans all 2049 characters of .S for a space and triggers MSan
diagnostic in AString::StrNCmp.