Correctly interpret status code of waitpid(), add test
WEXITSTATUS can be used only if WIFEXITED returns true, in fact
WEXITSTATUS returns 0 if the process was terminated by a signal. This
resulted in crashed tests being treated as passed by TestNode.

This change expands status checks to handle all state changes (normal
termination, termination by a signal, process receiving SIGSTOP).

Also added a test for TestNode that verifies that crash of a test
executable is detected and treated as test failure.