Replace plain loads and stores with atomic ones where necessary
In most cases plain loads and stores are replaced with atomic loads and
stores that use relaxed memory order. On x86 this results in the same
instructions being emitted, but allows ThreadSanitizer to know that
operation is intended to be atomic.

In one case (TCPConnectionPool::m_ThreadQuitNotification) relaxed memory
order wasn't enough, ThreadSanitizer reported data races in
TestDataTransfer test. So for this variable acquire/release memory order
is used.

In some places getters had to moved from a header to a .cpp file to
avoid including Atomic.h (which includes Windows.h) in the header.