Fix crash when struct variable is replaced by its member in Using()
When we are importing members of a struct into current scope via
function Using() we have to ensure that struct won't get deleted while
we iterating its members. This may happen if a struct has a member with
the same name as the struct itself.

This issue was found by BFFFuzzer, minimal reproducer is:
    .A=[.A=[]]Using(.A)