Allow a variable to be frozen multiple times
This effectively legalizes nested ForEach loops over the same array.
Previously there was an ASSERT that prevented such loops, but ASSERT
works only Debug builds and in Release such loops worked just fine in
most cases.

Issues were appearing only if the BFF code was trying to modify the
array after the end of a nested loop. In that case the array, which was
marked as not frozen at that point, was replaced without producing a
visible error. And later FunctionForEach::ParseFunction was accessing
already deleted variable on the second iteration.

This issue was found by BFFFuzzer, minimal(-ish) reproducer is:
    .a={'' 'X'}
    ForEach(.x in.a) {
        ForEach(.y in.a) {}
        ^a={}
    }