Support colored output from Clang and GCC
Both Clang and GCC (since 4.9) have support for colored diagnostic
output. Also both by default enable colored output when their stderr is
not redirected. This works well with other build drivers like make and
ninja because they don't intercept command output.

To make colored output work with FASTBuild we need to force it by
passing -fdiagnostics-color=always option to Clang/GCC when FASTBuild
output is not redirected.
Because not all GCC versions support -fdiagnostics-color options we
add that option only if -fdiagnostics-color=auto (which is the default)
is explicitly specified in .CompilerOptions.
Also we don't add that option if colored output is explicitly forced to
disabled or enabled in .CompilerOptions.