Fix null deference in MetaNone
MetaNone function is replaced with MetaNone type, this allows us to keep
the current syntax and specialize functions that take IMetaData& for
this new type. New overloads produce the same results as the old ones
did for a reference-to-null argument.
Despite the comment in the old MetaNone() implementation MetaNone()
wasn't used with operator+, so it doesn't need overloads for MetaNone
type.