Fix inconsistent parsing of function headers
Empty function headers (without any characters between parenthesis) were
incorrectly accepted for all functions, even for functions that require
headers.
For example, this code would correctly produce error #1001:
    ObjectList( )
    {
        .Compiler = '/usr/bin/gcc'
        .CompilerOptions = '-c %1 -o %2'
        .CompilerOutputPath = '.'
    }
And if the whitespace inside the function header would be removed then
there would be no error. Also in Debug build this would lead to a failed
assertion in Function::Commit because function name would be empty.

This change fixes the problem, but as a side effect specifying empty
parenthesis for functions that don't require headers (like Library) now
results in error #1001. This shouldn't be a problem because these empty
parenthesis could be just removed.

Also documentation for Compiler and ObjectList functions is updated to
match code behaviour.