Improve Test function logic
Display and save the test output when test times out. This allows to
determine which part of the test is most likely responsible for the
timeout.

Show "Test failed" message before trying to save the output because that
may fail as well and in that case the former message wouldn't be
displayed at all.