Fix bug where FBuild uses wrong database filename on Linux
FBuild.cpp was incorrectly checking if the __LINUX macro was defined. It should
have instead checked for the __LINUX__ macro. This resulted in a file called
"fbuild" being as the database file on Linux.

We happened to have an existing file with that name (the fbuild binary itself).
FBuild ended up reading our existing file and failng to parse it since it was
not the format it expected.