Fix race between processes in TestSharedMemory
It is possible that the child process created by fork() will try to open
shared memory object before it is created in the parent process. In that
case the child process dies and the test fails.
This can be reproduced reliably by inserting Thread::Sleep(1) before the
call to SharedMemory::Create.

To fix that we:
1. Create the shared memory object before forking.
2. Unmap its memory region to avoid inheriting it in the child process.
3. Open the shared memory again in both processes after forking.