Use correct extension for temporary files with preprocessed output
GCC decides how to interpret input files by looking at file's
extensions, in particular it determines the language of the input and
the fact whether input was already preprocessed or not. This applies to
"gcc" binary itself (not to g++, g++ "upgrades" .c files to C++).

The problem was that we always used ".i" extension for the temporary
files in which output of the preprocessor was stored, and that extension
tell GCC that the file is a preprocessed C program.
So if a C++ source was compiled with separate preprocessing step and
.Compiler referenced to a "gcc" binary then GCC interpreted that source
as C code and failed to compile it.

To fix that ObjectNode::WriteTmpFile now selects proper extension for
the temporary file based on the extension of the original file.