Update value of the "-x" option when compiling preprocessed source
It is possible to explicitly specify the type of an input file to GCC
and Clang via "-x" option.
Currently, when we are compiling a preprocessed source we don't touch
"-x" options and this forces GCC and Clang to preprocess the source a
second time because they are told so by the option.

To fix this problem we now change the value of the "-x" option to a
corresponding "cpp-output" variant.

This has 2 benefits:
1. Compilation of the preprocessed source is slightly faster.
2. During distributed compilation GCC running on the remote machine
   doesn't try to inject contents of /usr/include/stdc-predef.h from
   that machine into preprocessed source. This avoids potential problems
   that may arise when that file is different on the remote machine.