Check SO_ERROR to determine if connect() was successful on Linux/OSX
On Linux and OSX select() sets a flag in write fd_set both when
connect() succeeds and when it fails. To determine the result we must
get the value of SO_ERROR via getsockopt().

Before this change connection attempts to workers that were not running
were always treated as successful. For each such attempt a new thread
was created and then immediately terminated after a send() failure on
the non-connected socket.