Require header for CopyDir and RemoveDir
Header is the only source of the name for these nodes and thus it should be
required. Documentation already says that it is required but the code
doesn't enforce that.

Without the change this code is silently accepted in Release but
triggers assert about non-empty name in Function::Commit in Debug:
    CopyDir
    {
        .SourcePaths = 'src'
        .Dest = 'dst'
    }