Blacklist folder feature
Use a blacklist folder to exclude workers:
1. Users can create a hostname file in the blacklist folder in the brokerage path. Presence of this file
will exclude the worker from use by FBuild clients.
This is needed for workers that return error codes outside FBuild's known system error ranges. It allows clients to immediately (and manually) blacklist workers, not having to wait for system errors to be added to the FBuild and FBuildWorker exes. The user that creates the file can include whatever they want in the blacklist file contents, most likely the error message seen and their contact info.
2. This same feature also has an automatic behavior: When a FBuild client blacklists a worker in its known system error ranges, it will automatically create a blacklist file causing the worker to be blacklisted for other clients. The contents of the file will include the client machine name, client username, and the system error message seen.
3. If a blacklist file exists, then the FBuildWorker on the blacklisted machine will notice the file and show a popup dialog, notifying the worker's owner of the file path, so they can go and look at the error and fix their machine.