fbuild.exe with corrected symlink copy behavior
Previously, instead of the default behavior of following a link and copying the contents to a new file, CopyFile attempted to make a new link pointing at the same place.

But it did this by copying the link contents without modification, an a link's contents can be a relative path.  So if /a/b/foo was a symlink that contained simply "bar", then copying it to /a/d would result in a broken link (there is no /a/d/bar).

There may be situations where that is the desired behavior, but that really needs not to be the default.

This change makes the link target always be an absolute path, starting from /, so that it will work regardless of where the link is copied to.