make contents buffer a temporary variable
I was originally going to generate the contents during DetermineNeedToBuild so I could compare against the file on disk, but I couldn't come up with a reason why the output would change without fbuild already detecting the need to rebuild everything. So I moved generation to DoBuild but didn't remove the member variable.  Cleaning that up.