Use different token type for opening and closing brackets
All places in the code that check for a particular bracket type also
need to distinguish between opening and closing brackets.

In some places only the bracket type was checked and it resulted in code
like this to be parsed without errors:
    .Array = } 42 }
    .Struct = ]]

To simplify checks and to eliminate possibility of forgetting to check
for opening vs closing bracket, opening an closing bracket now use
different BFFTokenType values.