Only add include if preprocessor included it (#685)
When using tools like bison to generate a cpp file, it will leave `#line`
directives. This can cause fbuild to incorrectly add the file as a
dependency. Instead of always adding the included file, only add it if
the preprocessor included it. This can be detected by flags at the end of the
preprocessor line. Eg

```
cat main.cpp
 #include "found.h"
 #line 1 "notreal.h

touch found.h

g++ -E -c main.cpp
 # 1 "main.cpp"
 # 1 "<built-in>"
 # 1 "<command-line>"
 # 1 "/usr/include/stdc-predef.h" 1 3 4
 # 1 "<command-line>" 2
 # 1 "main.cpp"
 # 1 "found.h" 1
 # 2 "main.cpp" 2
 # 1 "notreal.h"
```

Lines which contain at least a ` 1` at the end should be the only files
tracked by fbuild since that is what the preprocessor includes.

https://gcc.gnu.org/onlinedocs/cpp/Preprocessor-Output.html

The reason why `GetSize()` is reduced by 1 is because the actual file
being compiled is in the preprocessed output and it now skipped (because
it does not have a ` 1` and fbuild tracks the file anyway)

Lines starting with `#line` for the clang fms-extension are still always
added

Fixes: https://github.com/fastbuild/fastbuild/issues/228