New function, TextFile, for creating text files (#629)
* New function, TextFile, for creating text files

The envisioned use of this function is the creation of response files with compiler parameters and text files with build system constants in it.  I'm sure there will be other uses once it's "in the wild."

* Command-line length is still an issue on windows, and the "solution" is to put the command-line arguments into a text file (for the tools that support it, such as cl.)  Until fastbuild is generating these "response files" automatically, the TextFile function will allow people to do that generation themselves.

* Some languages do not have a facility for passing define macros to the compiler, and the best equivalent is to generate a tiny source file with those constants in it.

* TextFile could be used to "touch" a file to trigger some other activity (such as an auto deploy.)

I didn't provide any facility to include existing data at this time because that's so much more complicated and it's not important to my use case.

* Cleaning up diffs

* make contents buffer a temporary variable

I was originally going to generate the contents during DetermineNeedToBuild so I could compare against the file on disk, but I couldn't come up with a reason why the output would change without fbuild already detecting the need to rebuild everything. So I moved generation to DoBuild but didn't remove the member variable.  Cleaning that up.

Co-authored-by: ffulin <franta.fulin@gmail.com>