TestProjectGeneration: Expect quoted paths in IntellisenseAndCodeSense test
This patch fixes the IntellisenseAndCodeSense test in case when the full
path to the repository checkout contains any character that forces
XCodeProjectGenerator to write quoted include paths to the .pbxproj file
(see XCodeProjectGenerator::ShouldQuoteString()).

Previously the test didn't account for possibility of a quoted path and
the trailing quote was left in paths saved to the includes array.