Fix Compress() and Decompress() in MultiBuffer (#845)
Calls to Replace() in Compress() and Decompress() in MultiBuffer worked
correctly only when Compressor::GetResultSize() from the second argument
was called before Compressor::ReleaseResult() from the first argument.

But the order of evaluation of function arguments is unspecified in C++,
so the code worked with some compilers (eg. GCC 10) and produced
incorrect results (zero data size in m_ReadStream or m_WriteStream) with
others (eg. Clang 12).

In the end this problem was caught by several tests:
* TestBuildFBuild.BuildCleanWithCache
* TestCache.ConsistentCacheKeysWithDist
* TestPrecompiledHeaders.TestPCHWithCache