Use MSVC headers in x64Clang configuration
This fixes build on Windows with Clang using VS2015 or VS2019 and new
WinSDK versions.
Currently the build fails with errors about not found include files
excpt.h and vcruntime.h, both are included in WinSDK headers:

This was observed with Clang 11 (installed via Visual Studio installer)
and WinSDK 10.0.19041.0.

We need to specify the include path to MSVC headers using /imsvc because
otherwise the build fails due to warnings in MSVC headers reported
Clang. So the .VSIncludePaths variable couldn't be reused.