TestSemaphore: Use larger timeout in WaitTimeout test
Revert timeouts (both for Semaphore::Wait() and for assert on
Timer::GetElapsed() result) to pre 14f5fa73d6 values (50ms and 25ms).

New values (1ms and 1ms) were too small and didn't provide any any
margin for error (despite the comment still claiming so). As a result
this test was failing sporadically on some machines (namely on Windows
runners for GitHub Actions).

Also upgrade ASSERT on Timer::GetElapsed() result to TEST_ASSERT, so it
would be checked in non-Debug builds.