Add CI via GitHub Actions (#848)
* Use MSVC headers in x64Clang configuration

This fixes build on Windows with Clang using VS2015 or VS2019 and new
WinSDK versions.
Currently the build fails with errors about not found include files
excpt.h and vcruntime.h, both are included in WinSDK headers:

This was observed with Clang 11 (installed via Visual Studio installer)
and WinSDK 10.0.19041.0.

We need to specify the include path to MSVC headers using /imsvc because
otherwise the build fails due to warnings in MSVC headers reported
Clang. So the .VSIncludePaths variable couldn't be reused.

* TestSemaphore: Use larger timeout in WaitTimeout test

Revert timeouts (both for Semaphore::Wait() and for assert on
Timer::GetElapsed() result) to pre 14f5fa73d6 values (50ms and 25ms).

New values (1ms and 1ms) were too small and didn't provide any any
margin for error (despite the comment still claiming so). As a result
this test was failing sporadically on some machines (namely on Windows
runners for GitHub Actions).

Also upgrade ASSERT on Timer::GetElapsed() result to TEST_ASSERT, so it
would be checked in non-Debug builds.

* Add CI via GitHub Actions

Currently the following jobs are defined:
* Linux (Build & Tests, GCC 10, Clang 12)
  Checks compilation with newest compilers (available out of the box),
  and runs tests (using Tests target).
* Linux (Build, GCC 7, Clang 9)
  Checks compilation with oldest compilers (available out of the box).
* Linux (ASan, GCC 10, Clang 12)
  Runs tests compiled with Clang in Address Sanitizer mode.
* Linux (MSan, GCC 10, Clang 12)
  Runs tests compiled with Clang in Memory Sanitizer mode.
* Linux (TSan, GCC 10, Clang 12)
  Runs tests compiled with Clang in Thread Sanitizer mode.
* OSX (Build & Tests)
  Checks compilation with Apple Clang 12 (the only version available out
  of the box), and runs tests (using Tests target).
* Windows (Build & Tests, VS 2019)
  Checks compilation with VS 2019 and Clang 11 and runs tests (using
  Tests target).
* Windows (Build, VS 2017)
  Checks compilation with VS 2017 and Clang 11.
* Windows (Build, VS 2015)
  Checks compilation with VS 2015 and Clang 11.
* Windows (Analyze, VS 2019)
  Builds All-x64-Analyze target with VS 2019 performing static analysis.