Avoid dependency on glibc 2.33 and 2.17 when building against them
Since the compatibility with old glibc versions was addressed previously
(b78492682) it has regressed several times:
* In 0.97 fbuild started requiring glibc 2.10 due to usage of accept4().
  That function first appeared in glibc 2.10 so this dependency is not
  easy to avoid. Also requiring glibc 2.10 shouldn't be a problem as no
  supported distribution is using glibc older than that.
* In 1.03 fbuild started requiring glibc 2.17 due to usage of
  clock_gettime(). Since that function was used well before 1.03, the
  dependency was most likely caused by building on a newer distribution
  than before.
* Very soon fbuild might start requiring glibc 2.33 due to usage of
  stat(), lstat() and fstat(), as the way they are implemented was
  changed in that version.

Fixing dependency on glibc 2.17 is easy - we use the __asm__(".symver")
trick to link to the old implementation.
According to glibc 2.17 release notes
(https://sourceware.org/legacy-ml/libc-announce/2012/msg00001.html)
the new version of clock_gettime() was introduced to improve performance
of singe-threaded programs, so we can continue to use the old version.

Fixing dependency on glibc 2.33 is harder as implementation of stat(),
lstat(), etc. was changed significantly.
Previously they were implemented as inline wrappers directly in
sys/stat.h and were calling internal implementation functions (like
__xstat()). And executables were then linking to those internal
functions. In 2.33 this indirection was removed and now executables are
linked to real stat(), lstat(), etc. in libc.so.6.
So to stay compatible with old glibc versions we have to implement those
wrappers ourselves when we are building with glibc 2.33 or above.