Avoid dependency on glibc 2.34 when building against it
Glibc 2.34 broke forward compatibility again in two entirely new ways.

The first issue is that in glibc 2.34 the way in which global
initialization and destruction functions are called was changed:
https://sourceware.org/git/?p=glibc.git;a=commitdiff;h=035c012e32c11e84d64905efaf55e74f704d3668

Previously each executable contained functions __libc_csu_init and
__libc_csu_fini that iterated through all global initialization and
destruction functions and the program entry point (_start) passed their
addresses to the main startup code (__libc_start_main) implemented in
libc.so.

In glibc 2.34 __libc_csu_init and __libc_csu_fini are removed, _start
passes null pointers to __libc_start_main and __libc_start_main finds
and executes global initialization and destruction functions by itself.
As this new behavior of _start is incompatible with __libc_start_main
present in old libc.so, __libc_start_main got a new symbol version and
now programs built with glibc 2.34 link to __libc_start_main@@GLIBC_2.34.

To be able to continue to use any glibc to build executables that will
run on older systems we undo all of that:
1. We implement our own __libc_csu_init and __libc_csu_fini when
   building with glibc >= 2.34.
2. We hook calls __libc_start_main by using --wrap= linker option.
3. In our hook for __libc_start_main we call the real old __libc_start_main.
4. If we are building with glibc >= 2.34 we pass addresses of our
   __libc_csu_* implementations to __libc_start_main.

The second issue is that in glibc 2.34 libpthread, libdl and librt were
merged into libc. Unversioned symlinks for these shared libraries are
no longer installed and instead dummy static libraries are provided.
This means that now we can't link with these libraries in a usual way
(-ldl, etc.), if we try we will instead link with dummy static
libraries, and as a result our binary won't load on system with older
glibc due to unresolved symbols.

To ensure that we are linked to all needed libraries we now link
directly to versioned shared library files.

Also many functions got new symbol versions, so we add corresponding
__asm__(".symver") lines to glibc_compat.h