Fix NodeGraph::CleanPath() for paths ending with \. or /. (#933)
Previously, in paths ending with \. or /. the trailing dot was not
removed. This made setting properties to the working directory
unintuitive.
For example setting .ProjectBasePath to '.' didn't work, strings in
VSProjectGenerator::m_BasePaths had a dot at the end, BeginsWithI()
check against absolute paths to project files never succeeded and all
files were dumped into the project without directory structure.
Of course setting .ProjectBasePath to './' did work, but that is not
very intuitive.