Add a clang-tidy compiler driver
Order of arguments is important in clang-tidy, it has two sets of them.
First one is directed at clang-tidy, second one is passed to the underlying
clang process. They are separated by a `--` argument.

Since we need to handle --config-file= (which belongs to the first set) in
a particular way, we need to be able to insert it at the start of the
command line arguments.