[Fix] Distributed compilation not working due to deadlocks under heavy network load
[Improvement] Move worker discovery to Client thread, reducing startup time
[Improvement] Parallelize connecting to workers
[Note] -dist will no longer be auto disabled if there are no workers at startup

When the Client received and responded to a packet, it would often do so on the receiving thread. Under heavy network load, the client could be blocked on this send, and because it was on the receive thread it could in turn block the remote sender causing a deadlock. Making matters even worse, the main Client thread could attempt to send a small status message, which would then also become blocked, effectively deadlocking all network connections and preventing all distributed compilation until a send timeout of 10 minutes.

To resolve this, the client now has a dedicated send thread for each connection in addition to the existing receive threads. Any send operations, whether from a receive thread or the main Client thread is enqueued for processing by this thread, eliminating the possibility of deadlocks.

As part of this refactor, there are two additional improvements:
 - The discovery of workers is moved to the Client thread, reducing startup time, particularly when using a brokerage path.
Connections to workers are done in parallel rather than serially, reducing the time for workers to be available, but also avoiding timeouts to certain workers stalling the connectivity to other workers.
 - The groundwork for another improvement is also put into place with this change. It will be possible for new workers to be discovered during a build, rather than discovery being performed only once at startup. To simplify this already large refactor, this has been left as a TODO, but  said refactor is designed with this extensibilty in mind.

Finally, the behavior of -dist when there are no workers is changed. Previously, distributed compilation would be disabled if no workers were discovered during startup. Since this is no longer performed at startup this is no longer possible. Additionally when workers can become discoverable during builds in the future, this also won't be desirable in many cases.